enwiki_cats_list_generator: generate a list of categories in enwiki
The actual module that generates a category list file that would be
consumed by the article list generator has been replaced with a sub-module
that generates a list of actual categories present in enwiki.

This would be used to produce the actual 'category list file' from the
given user input.