article_list_generator: correctly handle very long URI issue
When there are a huge number of relevant categories for a given user
query, trying to request petscan for articles which belong to all of
them would result in a "14 Very large request URI .." error.

Avoid this by splitting the list of categories with each split
containing a heuristically determined maximum number of categories
(using utils/index.js). Then the request is done for each split and
the results are merged and de-duped to form the article list.