article_list_generator: add a alternative way to generate article list
In case PetScan, the tool used to generate the article list for a given
set of categories, fails due to some reason we use the alternative
method of generating the article list using the MediaWiki API directly.

The only reason for avoiding the use of MW API  directly is that there
might be edge cases we have avoided that the toll would correctly
manage. Also, currently the article list generated doesn't contain other
articles found in the sub-categories of the categories of concern.