package_generator: add fallback to generate the offline package path
The fallback is used when the verbose output of mwoffliner is turned off
which results in no output to the stdout. In this case, we manually
generate the file name by noticing the pattern in which the file name is
generated. As this is not a reliable way we use this as fallback[1].

[1]: Note that parsing the stdout of mwoffliner (with verbose turned on)
isn't a reliable way either but this is worse than that.