article_list_generator: correctly request only for articles in main ns
Only the articles in the main namespace were required to generate the
offline package from the beginning of this project. The configuration
of PetScan was expected to be specified in this way. Unfortunate
and unnecessary percent encoding the corresponding configuration
line resulted in PetScan not identifying it as a valid configuration
and thus made it to return articles in all the namespaces.

So, fix the issue by not percent encoding the corresponding
configuration.