category_list_generator: change the way the user query is interpreted
Considering a category to be relevant if it contains any of the
elements in a user query resulted in a lot of categories being
identified as relevant.

So, consider the user query to be a comma separated list of keywords.
This lets us re-defining the set of relevant categories to be those
which contain every searchable element in a keyword.